Moved the `ensureSync` utility method from `test/mixed.js` into `test/helpers.js`.
